seprated out users and verifactions
