+++ lib/users/email.ex
@@ -1,4 +1,4 @@
-defmodule HubIdentityElixir.HubIdentity.Email do
+defmodule HubIdentityElixir.Users.Email do
   alias HubIdentityElixir.HubIdentity.Server
 
   def create(uid, address) do
